+++ services-queue/src/main/kotlin/com/yourssu/morupark/queue/business/ReadStatusResult.kt
@@ -1,0 +1,7 @@
+package com.yourssu.morupark.queue.business
+
+data class ReadStatusResult(
+    val status: TicketStatus,
+    val rank: Long,
+    val estimatedWaitingTime: Long
+)
