--- services-queue/src/main/kotlin/com/yourssu/morupark/queue/business/Status.kt
@@ -1,6 +1,0 @@
-package com.yourssu.morupark.queue.business
-
-enum class TicketStatus {
-    WAITING,
-    ALLOWED
-}
