+++ services-queue/src/main/kotlin/com/yourssu/morupark/queue/business/TicketStatus.kt
@@ -1,8 +1,6 @@
 package com.yourssu.morupark.queue.business
 
-data class ReadStatusResult(
-    val status: TicketStatus,
-    val rank: Long,
-    val totalAhead: Long,
-    val estimatedWaitingTime: Long
-)
+enum class TicketStatus {
+    WAITING,
+    ALLOWED
+}
